move all hook function annotypes to scanning.hooks
